refactor: :art: implemented more robust exit code system and compilation pipeline
[da1823] [rm1723]
Passing error codes as enums make it easier to identify the meaning of the codes. This also prevents invalid exit codes being used.

Changing the structure of the compile function from recursive matches to a sequential process makes it easier to scale the different stages of the compilation process. By passing the AST in and out at each step, AST-optimisation stages can be easily added without nesting more matches. It also shows the order of steps for the compiler better.